Split visualizer internals from visualizer.h
