+++ src/gate_visual.h
@@ -1,3 +1,10 @@
+/**
+ * @file   gate_visual.h
+ * @date   08/2020
+ * @author Tim van der Meer
+ * @brief  contains the parameters used to control a gate's visualization
+ */
+
 #ifndef QL_GATE_VISUAL_H
 #define QL_GATE_VISUAL_H
