+++ src/visualizer_internal.h
@@ -1,0 +1,74 @@
+/**
+ * @file   visualizer_internal.h
+ * @date   09/2020
+ * @author Tim van der Meer
+ * @brief  declaration of the visualizer's internals
+ */
+
+#ifndef QL_VISUALIZER_INTERNAL_H
+#define QL_VISUALIZER_INTERNAL_H
+
+#include "visualizer.h"
+#include "CImg.h"
+
+// These undefs are necessary to avoid name collisions between CImg and Lemon.
+#undef cimg_use_opencv
+#undef True
+#undef False
+#undef IN
+#undef OUT
+
+namespace ql
+{
+
+enum BitType {CLASSICAL, QUANTUM};
+
+struct Position4
+{
+	long x0;
+	long y0;
+	long x1;
+	long y1;
+};
+
+struct Position2
+{
+	long x;
+	long y;
+};
+
+struct NodePositionData
+{
+	const unsigned int gap;
+	const unsigned int labelColumnWidth;
+	const unsigned int cycleNumbersRowHeight;
+
+	const unsigned int column;
+	const unsigned int row;
+};
+
+struct CircuitData
+{
+	const unsigned int amountOfQubits;
+	const unsigned int amountOfClassicalBits;
+	const unsigned int amountOfCycles;
+};
+
+void validateLayout(const Layout layout);
+
+unsigned int calculateAmountOfBits(const std::vector<ql::gate*> gates, const std::vector<size_t> ql::gate::* operandType);
+unsigned int calculateAmountOfCycles(const std::vector<ql::gate*> gates);
+
+void drawCycleNumbers(cimg_library::CImg<unsigned char>& image, const Layout layout, const CircuitData circuitData);
+void drawBitLine(cimg_library::CImg<unsigned char>& image, const Layout layout, const BitType bitType, const unsigned int row, const CircuitData circuitData);
+void drawGroupedClassicalBitLine(cimg_library::CImg<unsigned char>& image, const Layout layout, const CircuitData circuitData);
+void drawGate(cimg_library::CImg<unsigned char>& image, const Layout layout, const CircuitData circuitData, ql::gate* const gate);
+
+void drawGateNode(cimg_library::CImg<unsigned char>& image, const Layout layout, const CircuitData circuitData, const Node node, const NodePositionData positionData);
+void drawControlNode(cimg_library::CImg<unsigned char>& image, const Layout layout, const CircuitData circuitData, const Node node, const NodePositionData positionData);
+void drawNotNode(cimg_library::CImg<unsigned char>& image, const Layout layout, const CircuitData circuitData, const Node node, const NodePositionData positionData);
+void drawCrossNode(cimg_library::CImg<unsigned char>& image, const Layout layout, const CircuitData circuitData, const Node node, const NodePositionData positionData);
+
+} // ql
+
+#endif //QL_VISUALIZER_INTERNAL_H
